move nav bar to its own folder
