+++ web/src/app/layout.tsx
@@ -3,7 +3,7 @@ import "./globals.css";
 import { ThemeProvider } from "@mui/material";
 import { theme } from '../theme/theme';
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter';
-import { NavBar } from "./components/NavBar";
+import { NavBar } from "./components/navBar/NavBar";
 
 export const metadata: Metadata = {
   title: "Representative Portal",
